obj: don't panic if a symbol has no data
It's useful to treat Data/Bounds as a generic "has data" interface,
but on Sym they currently panic for symbols with no data (e.g.,
undefined symbols), which makes them hard to use in a generic way
because caller may not know anything about undefined symbols.

Hence, replace the panic with an error return.